chore: rebrand to Ravel, strip Element CI, add Obtainium release workflow
