--- .github/workflows/fork-pr-notice.yml
@@ -1,37 +1,0 @@
-name: Community PR notice
-
-on:
-  workflow_dispatch:
-  pull_request_target: # zizmor: ignore[dangerous-triggers]
-    types:
-      - opened
-      - reopened
-
-permissions: {}
-
-jobs:
-  welcome:
-    runs-on: ubuntu-latest
-    permissions:
-      # Require to comment the PR.
-      pull-requests: write
-    name: Welcome comment
-    # Only display it if base repo (upstream) is different from HEAD repo (possibly a fork)
-    if: github.event.pull_request.base.repo.full_name != github.event.pull_request.head.repo.full_name
-    steps:
-      - name: Add auto-generated commit warning
-        uses: actions/github-script@ed597411d8f924073f98dfc5c65a23a2325f34cd # v8.0.0
-        with:
-          script: |
-            github.rest.issues.createComment({
-              issue_number: context.issue.number,
-              owner: context.repo.owner,
-              repo: context.repo.repo,
-              body: `Thank you for your contribution! Here are a few things to check in the PR to ensure it's reviewed as quickly as possible:
-            
-              - Your branch should be based on \`origin/develop\`, at least when it was created.
-              - The title of the PR will be used for release notes, so it needs to describe the change visible to the user.
-              - The test pass locally running \`./gradlew test\`.
-              - The code quality check suite pass locally running \`./gradlew runQualityChecks\`.
-              - If you modified anything related to the UI, including previews, you'll have to run the \`Record screenshots\` GH action in your forked repo: that will generate compatible new screenshots. However, given Github Actions limitations, **it will prevent the CI from running temporarily**, until you upload a new commit after that one. To do so, just pull the latest changes and push [an empty commit](https://coderwall.com/p/vkdekq/git-commit-allow-empty).`
-            })
