--- .github/workflows/post-release.yml
@@ -1,30 +1,0 @@
-name: Post-release
-
-on:
-  push:
-    tags:
-      - 'v*'
-
-permissions: {}
-
-jobs:
-  post-release:
-    runs-on: ubuntu-latest
-    # Skip in forks
-    if: github.repository == 'element-hq/element-x-android'
-
-    steps:
-      - name: Trigger pipeline
-        uses: actions/github-script@ed597411d8f924073f98dfc5c65a23a2325f34cd # v8.0.0
-        with:
-          github-token: ${{ secrets.ENTERPRISE_ACTIONS_TOKEN }}
-          script: |
-            const tag = context.ref.replace('refs/tags/', '');
-            const inputs = { git_tag: tag };
-            await github.rest.actions.createWorkflowDispatch({
-              owner: 'element-hq',
-              repo: 'element-enterprise',
-              workflow_id: 'pipeline-android.yml',
-              ref: 'main',
-              inputs: inputs
-            });
