--- .github/workflows/sync-localazy.yml
@@ -1,52 +1,0 @@
-name: Sync Localazy
-on:
-  workflow_dispatch:
-  schedule:
-    # At 00:00 on every Monday UTC
-    - cron: '0 0 * * 1'
-
-permissions: {}
-
-jobs:
-  sync-localazy:
-    runs-on: ubuntu-latest
-    # Skip in forks
-    if: ${{ github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == 'element-hq/element-x-android' }}
-    steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-        with:
-          persist-credentials: false
-      - name: Use JDK 21
-        uses: actions/setup-java@be666c2fcd27ec809703dec50e508c2fdc7f6654 # v5.2.0
-        with:
-          distribution: 'temurin' # See 'Supported distributions' for available options
-          java-version: '21'
-      - name: Configure gradle
-        uses: gradle/actions/setup-gradle@0723195856401067f7a2779048b490ace7a47d7c # v5.0.2
-        with:
-          cache-read-only: ${{ github.ref != 'refs/heads/develop' }}
-      - name: Set up Python 3.12
-        uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6.2.0
-        with:
-          python-version: 3.14
-      - name: Setup Localazy
-        run: |
-          curl -sS https://dist.localazy.com/debian/pubkey.gpg | sudo gpg --dearmor -o /etc/apt/trusted.gpg.d/localazy.gpg
-          echo "deb [arch=amd64 signed-by=/etc/apt/trusted.gpg.d/localazy.gpg] https://maven.localazy.com/repository/apt/ stable main" | sudo tee /etc/apt/sources.list.d/localazy.list
-          sudo apt-get update && sudo apt-get install localazy
-      - name: Run Localazy script
-        run: |
-          ./tools/localazy/downloadStrings.sh --all
-          ./tools/localazy/importSupportedLocalesFromLocalazy.py
-          ./tools/test/generateAllScreenshots.py
-      - name: Create Pull Request for Strings
-        uses: peter-evans/create-pull-request@c0f553fe549906ede9cf27b5156039d195d2ece0 # v8.1.0
-        with:
-          token: ${{ secrets.DANGER_GITHUB_API_TOKEN }}
-          commit-message: Sync Strings from Localazy
-          title: Sync Strings
-          body: |
-            - Update Strings from Localazy
-          branch: sync-localazy
-          base: develop
-          labels: PR-i18n
