--- .github/workflows/sync-sas-strings.yml
@@ -1,40 +1,0 @@
-name: Sync SAS strings
-on:
-  workflow_dispatch:
-  schedule:
-    # At 00:00 on every Monday UTC
-    - cron: '0 0 * * 1'
-
-permissions: {}
-
-jobs:
-  sync-sas-strings:
-    runs-on: ubuntu-latest
-    # Skip in forks
-    if: ${{ github.event_name != 'pull_request' || github.event.pull_request.head.repo.full_name == 'element-hq/element-x-android' }}
-    # No concurrency required, runs every time on a schedule.
-    steps:
-      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
-        with:
-          persist-credentials: false
-      - name: Set up Python 3.12
-        uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405 # v6.2.0
-        with:
-          python-version: 3.14
-      - name: Install Prerequisite dependencies
-        run: |
-          pip install requests
-      - name: Run SAS String script
-        run: ./tools/sas/import_sas_strings.py
-      - name: Create Pull Request for SAS Strings
-        uses: peter-evans/create-pull-request@c0f553fe549906ede9cf27b5156039d195d2ece0 # v8.1.0
-        with:
-          commit-message: Sync SAS Strings
-          title: Sync SAS Strings
-          body: |
-            - Update SAS Strings from matrix-doc.
-          branch: sync-sas-strings
-          base: develop
-          labels: PR-Misc
-
-
