--- .github/workflows/triage-incoming.yml
@@ -1,16 +1,0 @@
-name: Move new issues onto issue triage board v2
-
-on:
-  issues:
-    types: [ opened ]
-
-permissions: {}
-
-jobs:
-  triage-new-issues:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/add-to-project@244f685bbc3b7adfa8466e08b698b5577571133e # v1.0.2
-        with:
-          project-url: https://github.com/orgs/element-hq/projects/91
-          github-token: ${{ secrets.ELEMENT_BOT_TOKEN }}
