--- .github/workflows/validate-lfs.yml
@@ -1,15 +1,0 @@
-name: Validate Git LFS
-
-on: [pull_request, merge_group]
-
-permissions: {}
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    name: Validate
-    steps:
-      - uses: nschloe/action-cached-lfs-checkout@f46300cd8952454b9f0a21a3d133d4bd5684cfc2 # v1.2.3
-
-      - run: |
-          ./tools/git/validate_lfs.sh
