+++ app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         android:networkSecurityConfig="@xml/network_security_config"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.ElementX"
+        android:theme="@style/Theme.Ravel"
         tools:targetApi="33">
 
         <meta-data
@@ -59,7 +59,7 @@
             android:configChanges="orientation|screenSize|screenLayout|keyboardHidden|uiMode"
             android:exported="true"
             android:launchMode="singleTask"
-            android:theme="@style/Theme.ElementX.Splash"
+            android:theme="@style/Theme.Ravel.Splash"
             android:windowSoftInputMode="adjustResize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
