+++ app/src/main/kotlin/app/ravel/android/initializer/CacheCleanerInitializer.kt
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2025 Element Creations Ltd.
+ * Copyright 2022-2025 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package app.ravel.android.initializer
+
+import android.content.Context
+import androidx.startup.Initializer
+import io.element.android.features.cachecleaner.impl.CacheCleanerBindings
+import io.element.android.libraries.architecture.bindings
+
+class CacheCleanerInitializer : Initializer<Unit> {
+    override fun create(context: Context) {
+        context.bindings<CacheCleanerBindings>().cacheCleaner().clearCache()
+    }
+
+    override fun dependencies(): List<Class<out Initializer<*>>> = emptyList()
+}
