+++ app/src/main/kotlin/app/ravel/android/intent/DefaultIntentProvider.kt
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2025 Element Creations Ltd.
+ * Copyright 2023-2025 New Vector Ltd.
+ *
+ * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
+ * Please see LICENSE files in the repository root for full details.
+ */
+
+package app.ravel.android.intent
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import androidx.core.net.toUri
+import dev.zacsweers.metro.AppScope
+import dev.zacsweers.metro.ContributesBinding
+import io.element.android.libraries.deeplink.api.DeepLinkCreator
+import io.element.android.libraries.di.annotations.ApplicationContext
+import io.element.android.libraries.matrix.api.core.EventId
+import io.element.android.libraries.matrix.api.core.RoomId
+import io.element.android.libraries.matrix.api.core.SessionId
+import io.element.android.libraries.matrix.api.core.ThreadId
+import io.element.android.libraries.push.impl.intent.IntentProvider
+import app.ravel.android.MainActivity
+
+@ContributesBinding(AppScope::class)
+class DefaultIntentProvider(
+    @ApplicationContext private val context: Context,
+    private val deepLinkCreator: DeepLinkCreator,
+) : IntentProvider {
+    override fun getViewRoomIntent(
+        sessionId: SessionId,
+        roomId: RoomId?,
+        threadId: ThreadId?,
+        eventId: EventId?,
+        extras: Bundle?,
+    ): Intent {
+        return Intent(context, MainActivity::class.java).apply {
+            action = Intent.ACTION_VIEW
+            data = deepLinkCreator.create(sessionId, roomId, threadId, eventId).toUri()
+            extras?.let(::putExtras)
+        }
+    }
+}
