--- app/src/main/kotlin/io/element/android/x/ElementXApplication.kt
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2022-2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.x
-
-import android.app.Application
-import androidx.startup.AppInitializer
-import androidx.work.Configuration
-import dev.zacsweers.metro.createGraphFactory
-import io.element.android.libraries.di.DependencyInjectionGraphOwner
-import io.element.android.libraries.workmanager.api.di.MetroWorkerFactory
-import io.element.android.x.di.AppGraph
-import io.element.android.x.info.logApplicationInfo
-import io.element.android.x.initializer.CacheCleanerInitializer
-import io.element.android.x.initializer.CrashInitializer
-import io.element.android.x.initializer.PlatformInitializer
-
-class ElementXApplication : Application(), DependencyInjectionGraphOwner, Configuration.Provider {
-    override val graph: AppGraph = createGraphFactory<AppGraph.Factory>().create(this)
-
-    override val workManagerConfiguration: Configuration = Configuration.Builder()
-        .setWorkerFactory(MetroWorkerFactory(graph.workerProviders))
-        .build()
-
-    override fun onCreate() {
-        super.onCreate()
-        AppInitializer.getInstance(this).apply {
-            initializeComponent(CrashInitializer::class.java)
-            initializeComponent(PlatformInitializer::class.java)
-            initializeComponent(CacheCleanerInitializer::class.java)
-        }
-
-        logApplicationInfo(this)
-    }
-}
