--- app/src/main/kotlin/io/element/android/x/di/DefaultRoomGraphFactory.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2025 Element Creations Ltd.
- * Copyright 2023-2025 New Vector Ltd.
- *
- * SPDX-License-Identifier: AGPL-3.0-only OR LicenseRef-Element-Commercial.
- * Please see LICENSE files in the repository root for full details.
- */
-
-package io.element.android.x.di
-
-import dev.zacsweers.metro.ContributesBinding
-import io.element.android.appnav.di.RoomGraphFactory
-import io.element.android.libraries.di.SessionScope
-import io.element.android.libraries.matrix.api.room.JoinedRoom
-
-@ContributesBinding(SessionScope::class)
-class DefaultRoomGraphFactory(
-    private val sessionGraph: SessionGraph,
-) : RoomGraphFactory {
-    override fun create(room: JoinedRoom): Any {
-        return sessionGraph.roomGraphFactory
-            .create(room, room)
-    }
-}
