+++ app/src/main/res/values-night/themes.xml
@@ -7,11 +7,11 @@
   -->
 <resources>
 
-    <style name="Theme.ElementX.Splash" parent="Theme.SplashScreen">
+    <style name="Theme.Ravel.Splash" parent="Theme.SplashScreen">
         <item name="windowSplashScreenBackground">@color/splashscreen_bg_dark</item>
         <item name="windowSplashScreenAnimatedIcon">@drawable/transparent</item>
-        <item name="postSplashScreenTheme">@style/Theme.ElementX</item>
+        <item name="postSplashScreenTheme">@style/Theme.Ravel</item>
     </style>
 
-    <style name="Theme.ElementX" parent="Theme.Material3.Dark.NoActionBar" />
+    <style name="Theme.Ravel" parent="Theme.Material3.Dark.NoActionBar" />
 </resources>
