+++ app/src/main/res/values/themes.xml
@@ -6,12 +6,12 @@
   ~ Please see LICENSE files in the repository root for full details.
   -->
 <resources xmlns:tools="http://schemas.android.com/tools">
-    <style name="Theme.ElementX.Splash" parent="Theme.SplashScreen">
+    <style name="Theme.Ravel.Splash" parent="Theme.SplashScreen">
         <item name="windowSplashScreenBackground">@color/splashscreen_bg_light</item>
         <item name="windowSplashScreenAnimatedIcon">@drawable/transparent</item>
-        <item name="postSplashScreenTheme">@style/Theme.ElementX</item>
+        <item name="postSplashScreenTheme">@style/Theme.Ravel</item>
     </style>
-    <style name="Theme.ElementX" parent="Theme.Material3.Light.NoActionBar">
+    <style name="Theme.Ravel" parent="Theme.Material3.Light.NoActionBar">
         <item name="android:forceDarkAllowed" tools:targetApi="q">false</item>
     </style>
 </resources>
