+++ plugins/src/main/kotlin/config/BuildTimeConfig.kt
@@ -9,11 +9,11 @@
 package config
 
 object BuildTimeConfig {
-    const val APPLICATION_ID = "io.element.android.x"
-    const val APPLICATION_NAME = "Element X"
-    const val GOOGLE_APP_ID_RELEASE = "1:912726360885:android:d097de99a4c23d2700427c"
-    const val GOOGLE_APP_ID_DEBUG = "1:912726360885:android:def0a4e454042e9b00427c"
-    const val GOOGLE_APP_ID_NIGHTLY = "1:912726360885:android:e17435e0beb0303000427c"
+    const val APPLICATION_ID = "app.ravel.android"
+    const val APPLICATION_NAME = "Ravel"
+    const val GOOGLE_APP_ID_RELEASE = ""
+    const val GOOGLE_APP_ID_DEBUG = ""
+    const val GOOGLE_APP_ID_NIGHTLY = ""
 
     val METADATA_HOST_REVERSED: String? = null
     val URL_WEBSITE: String? = null
